+++ backend/src/socket_handlers/SocketConnection.js
@@ -1,10 +1,10 @@
-import BidHandler from './BidHandler.js';
+import BidSocketHandler from './BidSocketHandler.js';
 
 class SocketConnection {
   constructor(io) {
     if (!io) throw new Error('Socket.io instance required');
     this.io = io;
-    this.bidHandler = new BidHandler(io);
+    this.BidSocketHandler = new BidSocketHandler(io);
     this.initialize();
   }
 
@@ -12,17 +12,16 @@ class SocketConnection {
     this.io.on('connection', (socket) => {
       console.info(`Client connected: ${socket.id}`);
 
-      // When a user enters item route in client will enter items room in socket
       socket.on('join-item', (itemId) => {
-        this.bidHandler.handleJoinItemRoom(socket, itemId);
+        this.BidSocketHandler.handleJoinItemRoom(socket, itemId);
       });
 
       socket.on('place-bid', (data) => {
-        this.bidHandler.handleBidPlacement(socket, data);
+        this.BidSocketHandler.handleBidPlacement(socket, data);
       });
 
       socket.on('leave-item', (itemId) => {
-        this.bidHandler.handleLeaveItemRoom(socket, itemId);
+        this.BidSocketHandler.handleLeaveItemRoom(socket, itemId);
       });
 
       socket.on('disconnect', () => {
